requirements.txt
# Core runtime
hello-agents>=0.2.2,<0.3.0
openai>=1.0.0,<2.0.0
tiktoken>=0.12.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
tavily-python>=0.3.3,<1.0.0
rich>=13.7.0,<14.0.0

# Data and analysis stack
# Chosen to stay in stable wheel ranges for Python 3.10 on Windows / Conda.
numpy>=1.24.4,<2.0.0
pandas>=2.1.4,<2.3.0
scipy>=1.11.4,<1.13.0
matplotlib>=3.8.2,<3.9.0
seaborn>=0.13.2,<0.14.0
statsmodels>=0.14.1,<0.15.0
tabulate>=0.9.0,<1.0.0
xlrd>=2.0.1,<3.0.0
openpyxl>=3.1.0,<4.0.0
Pillow>=10.4.0,<11.0.0
pdfplumber>=0.11.0,<1.0.0
pypdf>=4.0.0,<5.0.0
reportlab>=4.0.0,<5.0.0
chromadb>=0.5.5,<1.0.0
PyYAML>=6.0.1,<7.0.0

# Notebook and demo environment
jupyter>=1.0.0,<2.0.0
notebook>=7.0.0,<8.0.0
gradio>=4.44.0,<5.0.0
huggingface_hub>=0.19.3,<1.0.0
